webproject/requirements.txt
asgiref==3.8.1
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.1
pytz==2023.3
requests==2.28.2
sqlparse==0.5.0
tzdata==2023.3
